Add misc scripts and BIP39 hack
Currently works with 12-words-long mnemonics. Can fix the last word
of the Electrum mnemonic to work with any BIP39. Of course it does
not lead to the same keys, but one may find it handy if they remember
the words by heart already and want to use them with BIP39 software
like Aqua Wallet or Blockstream Green.

Example usage and output:

    $ ./rxf.sh \
      circle turkey border bacon immune distance \
      enjoy uniform base myth cheap news \
      | sh bintohex.sh | ./xf.sh
    circle turkey border bacon immune distance enjoy uniform base myth cheap news
    1 (hex) = 0001 (bin)
    001010010011110101010100011001110000100010100111000110100111111101010010101011110110101100010010111100100101000010011011110010100001
    circle turkey border bacon immune distance enjoy uniform base myth cheap neither